+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Pass `inflate` / `deflate` options to `up` / `down` functions
 * Add `Serializable`, an abstract base class for serializables
 * Async serialization/deserialization (`deflate.resolve` and `inflate.resolve`)
 * Serialization/deserialization to/from a POJO (`as` and `toPojo` options)